refactor(_utils): split lib.rs into manifest/project/dylint modules
Three focused submodules, each with one job:
- \`manifest\`: TOML structs + \`fixture_cargo_toml\` +
  \`fixture_dylint_toml\` (build the Rust values, serialise them).
- \`project\`: \`build_project\` (write the fixture to disk).
- \`dylint\`: \`run_dylint\` (shell out to \`cargo dylint\`).

The orchestrator \`run_project_with_sources\` stays at the crate
root since it spans all three. Everything is \`pub use\`-d at the
root so existing call sites in tests/_utils.rs and the warmup bin
keep their current imports.
