+++ utils/src/dylint.rs
@@ -1,0 +1,22 @@
+//! Shell out to `cargo dylint --all` and capture its output.
+
+use std::{path::Path, process::Command};
+
+use command_extra::CommandExtra;
+use pipe_trait::Pipe;
+
+/// Run `cargo dylint --all` inside `project_dir`, with
+/// `CARGO_TARGET_DIR` pointed at `shared_target_dir` so the build
+/// artefacts are reused across invocations.
+pub fn run_dylint(project_dir: &Path, shared_target_dir: &Path) -> (String, bool) {
+    let output = "cargo"
+        .pipe(Command::new)
+        .with_arg("dylint")
+        .with_arg("--all")
+        .with_current_dir(project_dir)
+        .with_env("CARGO_TARGET_DIR", shared_target_dir)
+        .output()
+        .expect("failed to run `cargo dylint`");
+    let stderr = String::from_utf8(output.stderr).expect("dylint stderr is not UTF-8");
+    (stderr, output.status.success())
+}
